New Country and Locale classes
The new Country and Locale classes allow you to manage languages and localization in a more adherent way to the standards. The sample project Pair_example has been contextually updated and contains a module (downloadable as a ZIP package) that allows you to adapt a project created for the older versions of Pair to the new version 1.4.